Revert "NVIDIA: VR: SAUCE: cxl: Move HDM decoder and register map definitions to include/cxl/cxl.h"
BugLink: https://bugs.launchpad.net/bugs/2163076

This reverts commit c9c6c8e67018c41531fca8f142d29818aefa8bff.

Acked-by: Seth Forshee <sforshee@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
